[SYCL] Insert SYCLKernelCallStmt and OutlinedFunctionDecl AST nodes for sycl_kernel_entry_point attributed functions.
Functions declared with the sycl_kernel_entry_point attribute are used to
generate SYCL kernel entry points. These functions represent the host vs
device split with a SYCLKernelCallStmt as the body of the function. That
statement represents the body of the function as written in its associated
original statement and an outline of a device-side function with a
translated body in its associated OutlinedFunctionDecl. This change adds
the SYCLKernelCallStmt and OutlinedFunctionDecl AST nodes for such functions.

 Conflicts:
	clang/lib/Sema/SemaDecl.cpp
	clang/lib/Sema/SemaSYCL.cpp