[SYCL] Initial support for reflecting SYCL kernel names and other properties.
Details of SYCL kernels, including the name of the generated SYCL kernel caller
function, must be coordinated across host and device compilations and made
available to the SYCL run-time library. This change introduces the initial
support for collecting such details and storing them within the ASTContext
for exposure via builtin functions to be added in a future change.

As part of this change, proper naming for SYCL kernel caller functions is
introduced. These generated functions are given names equivalent to a function
template specialization of a '__sycl_kernel_caller' function template that is
parameterized by a single type template parameter and has a void return type
and no function parameters. At present, these functions are always given
names that are encoded according to the Itanium ABI, but this might change in
the future.