[SYCL][NFC] Require a canonical type for kernel registration.
SYCL kernels are indexed by a (potentially qualified) type that uniquely
identifies a kernel. Previously, a QualType was used for the type index, but
this could result in a lookup failure when the stored type and the lookup
type differed with regard to type sugar. With this change, a canonical type
is required to both store and lookup a SYCL kernel.