[SYCL] Correct DeclRefExpr substitution in generated OutlinedFunctionDecl AST.
Previously, the DeclRefExpr substitution performed when generating the
OutlinedFunctionDecl AST created the replacement DeclRefExpr AST node with
the type of the VarDecl node being substituted for the ParmVarDecl node.
This was wrong because it didn't take into account adjustment of a reference
type to a non-reference type for the expression result. Further, the code
lacked checks that the substituting VarDecl type matched the ParmVarDecl
type and was not more cv-qualified.