[SYCL] Enable the sycl_kernel_entry_point attribute for SYCL host compilation.
Previously, the sycl_kernel_entry_point attribute was only recognized during
SYCL device compilation. This change adds recognition for host compilation so
that SYCL kernel metadata is generated for such attributed functions as needed
to support builtin functions like __builtin_sycl_kernel_name().

SYCLKernelCallStmt is now handled as a no-op during code generation. This
suffices to ensure that the body of a sycl_kernel_entry_point attributed
function is never evaluated during host execution; the body of such a function
holds a pattern for device execution and is solely used in the generation of
SYCL kernel caller functions. Invocation of a sycl_kernel_entry_point attributed
function in a host context is only intended to trigger emission of a SYCL kernel
caller function and to generate associated SYCL kernel metadata.

 Conflicts:
	clang/include/clang/Basic/Attr.td