[SYCL] Implement builtins
Implemented support for __builtin_sycl_kernel_param_kind,
__builtin_sycl_kernel_param_size and __builtin_sycl_kernel_param_offset

Changes cause ast-dump-sycl-kernel-call-stmt.cpp to fail. The failure
occurs when we try to obtain the size of a parameter with an
instantiated type. The failure is because the definition of the type
is not instantiated when attempting to obtain it's size. It is
unclear why this is happening. Maybe KernelInfo is being collected
too early? Further debugging is required.