[SYCL][Runtime] Do not pass kernel object by reference.
In the new design, the kernel object is passed as whole to
the device if it does not contain special types. The
type of parameter of the kernel argument is determined by
the parameter to the entry point function. A reference type
does not work in the current memory model.